Added easyzlib to the project so the data can be compressed, which should result in much smaller fpreplay files.
added a little sleep before displaying text that the fp data will be recorded, maybe helps to prevent crashes. (ugly needs better implementation)

changed the replay save function so it wont be the second last replay anymore (what a stupid shit)

added zeros to the fpreps files so they will be in order.

delete the lastfpreplay.rep as its not really a fp replay.

added function to cope with broodwar retarded calling conventions.

the data is now read as binary instead of text (FUCK WINDOWS)